+++ Interpreter/src/test/scala/org/enso/interpreter/semantic/CurryingTest.scala
@@ -1,4 +1,4 @@
-package org.enso.interpreter;
+package org.enso.interpreter.semantic
 
 class CurryingTest extends LanguageTest {
   "Functions" should "allow partial application" in {
